rename: 폴더명 변경
model -> domain
lib -> util
