+++ src/domain/GameManager.js
